+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to `Wildfire` will be documented in this file.
 
 ### Added
 - Traits
+- `table` for manually specifying the table associated with the Model
 
 ### Fixed
 - Issue in getting foreign tables